+++ apps/mobile/src/hooks/use-subscription.ts
@@ -1,7 +1,6 @@
 import {
   useQuery,
   useMutation,
-  useQueryClient,
   type UseQueryResult,
   type UseMutationResult,
 } from '@tanstack/react-query';
@@ -11,10 +10,6 @@ import type {
   Subscription,
   Usage,
   FamilySubscription,
-  CheckoutRequest,
-  CheckoutResponse,
-  CancelResponse,
-  PortalResponse,
 } from '@eduagent/schemas';
 import { useApiClient } from '../lib/api-client';
 import { useProfile } from '../lib/profile';
@@ -31,14 +26,6 @@ export type UsageData = Usage;
 
 /** Matches the warningLevel enum from @eduagent/schemas usageSchema */
 export type WarningLevel = 'none' | 'soft' | 'hard' | 'exceeded';
-
-interface TopUpResult {
-  topUp: {
-    amount: number;
-    clientSecret: string;
-    paymentIntentId: string;
-  };
-}
 
 interface ByokWaitlistResult {
   message: string;
@@ -162,99 +149,6 @@ export function useSubscriptionStatus(): UseQueryResult<SubscriptionStatusData> 
     },
     enabled: !!activeProfile,
     staleTime: 60_000, // 1 min — fast endpoint, no need for aggressive refetching
-  });
-}
-
-// ---------------------------------------------------------------------------
-// Mutation hooks — Stripe (web client)
-// ---------------------------------------------------------------------------
-// Kept for future web client — not used by mobile IAP flow.
-// Mobile billing uses native IAP via RevenueCat (see use-revenuecat.ts).
-// These hooks call Stripe checkout/portal/cancel API routes which are
-// dormant for mobile but will be activated when a web client is added.
-// ---------------------------------------------------------------------------
-
-/** Kept for future web client — not used by mobile IAP flow. */
-export function useCreateCheckout(): UseMutationResult<
-  CheckoutResponse,
-  Error,
-  CheckoutRequest
-> {
-  const client = useApiClient();
-
-  return useMutation({
-    mutationFn: async (input: CheckoutRequest): Promise<CheckoutResponse> => {
-      const res = await client.subscription.checkout.$post({ json: input });
-      await assertOk(res);
-      return (await res.json()) as CheckoutResponse;
-    },
-  });
-}
-
-/** Kept for future web client — not used by mobile IAP flow. */
-export function useCancelSubscription(): UseMutationResult<
-  CancelResponse,
-  Error,
-  void
-> {
-  const client = useApiClient();
-  const queryClient = useQueryClient();
-  const { activeProfile } = useProfile();
-
-  return useMutation({
-    mutationFn: async (): Promise<CancelResponse> => {
-      const res = await client.subscription.cancel.$post({ json: {} });
-      await assertOk(res);
-      return (await res.json()) as CancelResponse;
-    },
-    onSuccess: () => {
-      void queryClient.invalidateQueries({
-        queryKey: ['subscription', activeProfile?.id],
-      });
-    },
-  });
-}
-
-/** Kept for future web client — not used by mobile IAP flow. */
-export function useCreatePortalSession(): UseMutationResult<
-  PortalResponse,
-  Error,
-  void
-> {
-  const client = useApiClient();
-
-  return useMutation({
-    mutationFn: async (): Promise<PortalResponse> => {
-      const res = await client.subscription.portal.$post({ json: {} });
-      await assertOk(res);
-      return (await res.json()) as PortalResponse;
-    },
-  });
-}
-
-/** Kept for future web client — not used by mobile IAP flow. */
-export function usePurchaseTopUp(): UseMutationResult<
-  TopUpResult,
-  Error,
-  void
-> {
-  const client = useApiClient();
-  const queryClient = useQueryClient();
-  const { activeProfile } = useProfile();
-
-  return useMutation({
-    mutationFn: async (): Promise<TopUpResult> => {
-      const res = await client.subscription['top-up'].$post({
-        json: { amount: 500 },
-      });
-      await assertOk(res);
-      return (await res.json()) as TopUpResult;
-    },
-    onSuccess: () => {
-      void queryClient.invalidateQueries({
-        queryKey: ['usage', activeProfile?.id],
-      });
-    },
   });
 }
